MAYBE THE BEST WORKING GAME EXAMPLE
I just fixed the UI thread overloading problem. The handler was in a while loop always posting the TextView with some text. So I fixed that by checking it if it just posts once. In my previous commit I used total of 3 threads, but now I've come back to just 2(UI and myThread). So Its All working :D